PostgreSQL OIDs overflow fix
Former-commit-id: a9e330b1cfdc7d7302ac17ac3df4ee71fcfbda6b

+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/session/PostgreSessionManager.java
@@ -75,7 +75,7 @@ public class PostgreSessionManager implements DBAServerSessionManager<PostgreSes
     {
         try {
             try (JDBCPreparedStatement dbStat = ((JDBCSession) session).prepareStatement("SELECT pg_catalog.pg_terminate_backend(?)")) {
-                dbStat.setInt(1, sessionType.getPid());
+                dbStat.setLong(1, sessionType.getPid());
                 dbStat.execute();
             }
         }
